fix: Rust logging can be set by SetLogLevel and environment variables
