+++ message/v3/asynchronous_message.go
@@ -148,7 +148,7 @@ func NewAsynchronousPact(config Config) (*AsynchronousPact, error) {
 		return nil, err
 	}
 
-	native.Init()
+	native.Init(string(logging.LogLevel()))
 
 	return provider, err
 }
