+++ message/v4/asynchronous_message.go
@@ -222,7 +222,7 @@ func NewAsynchronousPact(config Config) (*AsynchronousPact, error) {
 		return nil, err
 	}
 
-	native.Init()
+	native.Init(string(logging.LogLevel()))
 
 	return provider, err
 }
